refactor: move zip
