+++ packages/files/outer-layout.html
@@ -73,8 +73,8 @@
               icon: "apps",
             },
             {
-              name: "系统文件",
-              path: "system",
+              name: "依赖包",
+              path: "packages",
               icon: "safe",
             },
           ],
